Merge 28 to main

+++ modules/media/ajax/FileUpload.php
@@ -36,10 +36,6 @@ function editFile()
 {
     $db   = \NDB_Factory::singleton()->database();
     $user =& User::singleton();
-    if (!$user->hasPermission('media_write')) {
-        showMediaError("Permission Denied", 403);
-        exit(0);
-    }
 
     // Read JSON from STDIN
     $stdin = file_get_contents('php://input');
@@ -51,6 +47,27 @@ function editFile()
 
     if (!$idMediaFile) {
         showMediaError("Media ID $idMediaFile not found", 404);
+        exit(0);
+    }
+
+    $row = $db->pselectRow(
+        "SELECT s.CenterID FROM media m
+         JOIN session s ON m.session_id = s.ID
+         WHERE m.id = :id",
+        ['id' => $idMediaFile]
+    );
+
+    if (!$row) {
+        showMediaError("Media ID $idMediaFile not found", 404);
+        exit(0);
+    }
+
+    if (!$user->hasPermission('media_write')
+        || (!$user->hasPermission('access_all_profiles')
+        && !$user->hasCenter(new \CenterID(strval($row['CenterID']))))
+    ) {
+        showMediaError("Permission Denied", 403);
+        exit(0);
     }
 
     $dateTaken = $req['dateTaken'];
@@ -188,6 +205,12 @@ function uploadFile()
     );
     $dstfile   = \Utility::appendForwardSlash($mediaPath)
         . \Utility::resolvePath($fileName);
+
+    if (file_exists($dstfile)) {
+        showMediaError("Conflict", 409);
+        return;
+    }
+
     if (move_uploaded_file($_FILES["file"]["tmp_name"], $dstfile)) {
         try {
             // Insert or override db record if file_name already exists
@@ -378,7 +401,6 @@ function getUploadFields()
         'visits'      => $visitList,
         'instruments' => $instrumentsList,
         'mediaData'   => $mediaData,
-        'mediaFiles'  => array_values(getFilesList()),
         'sessionData' => $sessionData,
         'language'    => $languageList,
         'startYear'   => $startYear,
@@ -431,25 +453,6 @@ function toSelect($options, $item, $item2)
     }
 
     return $selectOptions;
-}
-
-/**
- * Returns an array of (id, file_name) pairs from media table
- *
- * @return array
- * @throws DatabaseException
- */
-function getFilesList()
-{
-    $db       = \NDB_Factory::singleton()->database();
-    $fileList = $db->pselect("SELECT id, file_name FROM media", []);
-
-    $mediaFiles = [];
-    foreach ($fileList as $row) {
-        $mediaFiles[$row['id']] = $row['file_name'];
-    }
-
-    return $mediaFiles;
 }
 
 /**
